♻️ refactor: Home -> HomeUser 등 세분화
