+++ app/src/pages/HomePage/contents/Coalition/CoalitionScoreDynamic.tsx
