+++ app/src/pages/HomePage/contents/Eval/AverageCommentLength.tsx
