+++ app/src/pages/HomePage/contents/Eval/CurrWeekAverageEvalCount.tsx
