+++ app/src/pages/HomePage/contents/Eval/index.ts
@@ -1,0 +1,4 @@
+export * from './AverageCommentLength';
+export * from './AverageFeedbackLength';
+export * from './CurrWeekAverageEvalCount';
+export * from './TotalEvalCount';
