+++ app/src/pages/HomePage/contents/Team/CurrRegisteredCountRank.tsx
