+++ app/src/pages/HomePage/contents/Team/index.ts
@@ -1,0 +1,2 @@
+export * from './CurrRegisteredCountRank';
+export * from './LastExamResult';
