+++ app/src/pages/HomePage/contents/User/BlackholedPercentage.tsx
