+++ app/src/pages/HomePage/contents/User/CorrectionPointRanks.tsx
