+++ app/src/pages/HomePage/contents/User/MemberPercentage.tsx
