+++ app/src/pages/HomePage/contents/User/WalletRanks.tsx
