+++ app/src/pages/HomePage/contents/User/index.ts
@@ -1,20 +1,11 @@
 export * from './ActiveUserCountRecords';
 export * from './AverageCircleDurations';
-export * from './AverageCommentLength';
-export * from './AverageFeedbackLength';
 export * from './BlackholedCountPerCircles';
 export * from './BlackholedPercentage';
-export * from './CoalitionScoreDynamic';
-export * from './CoalitionScoreSum';
 export * from './CorrectionPointRanks';
 export * from './CurrMonthBlackholedCount';
-export * from './CurrRegisteredCountRank';
-export * from './CurrWeekAverageEvalCount';
 export * from './CurrWeekEvalCount';
 export * from './Hero';
-export * from './LastExamResult';
 export * from './MemberPercentage';
-export * from './TigCountPerCoalition';
-export * from './TotalEvalCount';
 export * from './UserCountPerLevels';
 export * from './WalletRanks';
